requirements.txt
python-dotenv==1.0.0
python-bitcoinlib==0.11.0
requests==2.31.0
